+++ src/Phema.RabbitMq.Consumers/Queues/Extensions/RabbitMQQueueBuilderExtensions.cs
@@ -139,7 +139,8 @@ namespace Phema.RabbitMQ
 			this IRabbitMQQueueBuilder configuration,
 			byte priority)
 		{
-			return configuration.WithArgument("x-max-priority", priority);
+			// Hack, because RabbitMQ.Client has no conversion to byte
+			return configuration.WithArgument("x-max-priority", (int)priority);
 		}
 
 		/// <summary>
